added create template that works with pusher when posting work order
